requirements.txt
pyEDFlib>=0.1.33
numpy~=1.26.1
pandas~=2.1.2
beautifulsoup4~=4.12.2
lxml
tqdm~=4.66.1
PyYAML~=6.0.1
matplotlib~=3.8.1
scikit-learn~=1.3.2
yaml~=0.2.5
torchinfo~=1.8.0
sortedcontainers~=2.4.0